==> OvarianCarcinosarcoma/notebooks/S000322/1C_assignConfidTable.Rmd <==
---
title: "Assignment confidence table"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float:
      smooth_scroll: true
    theme: readable
    highlight: tango 
    df_print: paged
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      dev = "png",
                      dpi = 300,
                      fig.asp=0.6, 
                      fig.path='./figures/assign_table/',
                      warning=FALSE, message=FALSE)
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = here::here())
library(DropletUtils)
library(tidyverse)
library(patchwork)
library(scuttle)
source(here::here("scripts/R/theme_publication.R"))
theme_set(theme_Publication())
```

# Aim
I previous compared Total-Seq A hashtag, MULTI-Seq and CellPlex lipid on human PBMCs. All protocols worked pretty well but human PBMCs are the easiest sample to work with.

Then I tried mouse embryonic brain E18.5 which didn't work that well. CellPlex diluted 1:10 worked best here.

## This sample
Run a more challenging sample = Ovarian carcinosarcome PDX nuclei.  
Four PDXs were processed with 10x Genomics nuclei isolation kit. A capture was immdiately performed which is the unlabelled.  
Three more aliquots were labelled in singlicate with cell multiplexing reagents and captured later.

## This analysis
Cell ranger multi did not demultiplex cells at the default range of 99%. Investigate further by manually adjusting the confidence table.

## Comparisons

* Fresh unlabelled nuclei after nuclei isolation with 10x Genomics kit
* Fixed nuclei nuclei isolation with 10x Genomics kit
* Fresh nuclei CellPlex 1:10 diluted
* Fresh nuclei custom MULTI-Seq CMO (Nextera mod)
* Fresh nuclei 1ug nucleus hashtag Mab414

# Load Data

Find the paths to the cell ranger assignment confidence table.

```{r dataPaths}

cmo_path <- '/stornext/Projects/score/GenomicsRnD/DB/S000322/cellranger_outs/G000225_PDX_CMO/multi/multiplexing_analysis/'
hto_path <- '/stornext/Projects/score/GenomicsRnD/DB/S000322/cellranger_outs/G000225_PDX_HTO/multi/multiplexing_analysis/'
nxt_path <- '/stornext/Projects/score/GenomicsRnD/DB/S000322/cellranger_outs/G000225_PDX_NXT/multi/multiplexing_analysis/'

```

Load data using DropletUtils

```{r loadData}
cmo_table <- read_csv(paste(cmo_path, "assignment_confidence_table.csv", sep=""))
hto_table <- read_csv(paste(hto_path, "assignment_confidence_table.csv", sep=""))
nxt_table <- read_csv(paste(nxt_path, "assignment_confidence_table.csv", sep=""))
```

# Visualise the assignment probability {.tabset}

## CellPlex 1:10 dilution

Consistent with the wetting failure and general poor performance CellPlex was unsatisfactory.

```{r cmo_confidence}
plt1 <- ggplot(cmo_table, 
             aes(x = Assignment_Probability)) + 
  geom_density() + 
  xlab("Assignment probability") + 
  ylab("Frequency") +
  ggtitle("CellPlex")

plt1
```

## Hashtag antibody

The only protocol with decent distribution of assignment cconfident. The natural threshoold appears around 0.7.

```{r hto_confidence}
plt2 <- ggplot(hto_table, 
             aes(x = Assignment_Probability)) + 
  geom_density() + 
  xlab("Assignment probability") + 
  ylab("Frequency") +
  ggtitle("Hashtag Ab")

plt2
```

## MULTI-Seq CMO

Poor performance.

```{r nxt_confidence}
plt3 <- ggplot(nxt_table, 
             aes(x = Assignment_Probability)) + 
  geom_density() + 
  xlab("Assignment probability") + 
  ylab("Frequency") +
  ggtitle("MULTI-Seq CMO")

plt3
```

## Group together

```{r group, fig.asp=1}
plt1 / plt2/ plt3
```

# Session Info

```{r}
sessionInfo()
```
