==> OvarianCarcinosarcoma/notebooks/S000322/2C_GEX_cluster_visualise.Rmd <==
---
title: "GEX Clustering and dimension reduction"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float:
      smooth_scroll: true
    theme: readable
    highlight: tango 
    df_print: paged
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      dev = "png",
                      dpi = 300,
                      fig.asp=1, 
                      fig.path='./figures/gex_clustering/',
                      warning=FALSE, message=FALSE)
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = here::here())
suppressPackageStartupMessages({
  library(tidyverse)
  library(patchwork)
  library(scuttle)
  library(scater)
  library(scran)
  })

source(here::here("scripts/R/theme_publication.R"))
theme_set(theme_Publication(base_size = 18))
```

# Aim

Visulaise the dimension reduction results with respect to the metadata.

## Samples

* Fresh nuclei CellPlex 1:10 diluted
* Fresh nuclei custom MULTI-Seq CMO (Nextera mod)
* Fresh nuclei 1ug nucleus hashtag Mab414
* Fresh unlabelled nuclei after nuclei isolation with 10x Genomics kit
* Fresh unlabelled nuclei with ambient RNA removal
# Load Data

From the 2B_QC_dimred notebook

```{r loadData}
sce_list <- readRDS((
  "/vast/scratch/users/brown.d/S000322/S000322.dimred.rds"
))

capture <- names(sce_list)
capture
```

# Clustering

We build a shared nearest neighbour graph and use the Louvain algorithm to identify clusters. We build the graph using the principal components.

## Initial clustering

Smaller 'k' usually yields finer clusters.

Clustering is stable from k=5 to k=10 with 5 clusters resolved for the higher quality datasets

```{r cluster_1}
set.seed(911)

for (n in names(sce_list)) {
  current <- sce_list[[n]]
  snn_gr_cmo <- buildSNNGraph(current, use.dimred = "PCA", k = 7) 
  clusters <- igraph::cluster_louvain(snn_gr_cmo)
  current$cluster <- factor(clusters$membership)
  sce_list[[n]] <- current
}

```

# Visualise UMAP by QC parameters {.tabset}

## Cluster number

```{r umap_cluster, fig.height=8, fig.width=10}
umap_plt <- list()

for (i in 1:length(sce_list)) {
  current <- sce_list[[i]]
  umap_plt[[i]] <- plotReducedDim(current, dimred="UMAP", colour_by="cluster") + 
    ggtitle(capture[i]) + theme_Publication(base_size = 18)
}

umap_plt[[4]] + umap_plt[[1]] + umap_plt[[2]] + umap_plt[[3]]
```

## The SNP donor

```{r umap_donor, fig.height=8, fig.width=10}
umap_plt <- list()

for (i in 1:length(sce_list)) {
  current <- sce_list[[i]]
  umap_plt[[i]] <- plotReducedDim(current, dimred="UMAP", colour_by="donor_id") + 
    ggtitle(capture[i]) + theme_Publication(base_size = 18)
}

umap_plt[[4]] + umap_plt[[1]] + umap_plt[[2]] + umap_plt[[3]]
```

## Library size

```{r umap_libSize, fig.height=8, fig.width=10}
umap_plt <- list()

for (i in 1:length(sce_list)) {
  current <- sce_list[[i]]
  current$logSum <- log10(current$sum +1)
  umap_plt[[i]] <- plotReducedDim(current, dimred="UMAP", colour_by="logSum") + 
    ggtitle(capture[i]) + theme_Publication()
}

umap_plt[[1]] + umap_plt[[2]] + umap_plt[[3]] +
  umap_plt[[4]] 
```

## Mouse percentage

```{r umap_mouse, fig.height=8, fig.width=10}
umap_plt <- list()

for (i in 1:length(sce_list)) {
  current <- sce_list[[i]]
  umap_plt[[i]] <- plotReducedDim(current, dimred="UMAP", colour_by="altexps_Mouse_percent") + 
    ggtitle(capture[i]) + theme_Publication()
}

umap_plt[[1]] + umap_plt[[2]] + umap_plt[[3]] +
  umap_plt[[4]]
```

## Mitochondrial percentage

```{r umap_mito, fig.height=8, fig.width=10}
umap_plt <- list()

for (i in 1:length(sce_list)) {
  current <- sce_list[[i]]
  umap_plt[[i]] <- plotReducedDim(current, dimred="UMAP", colour_by="subsets_Mito_percent") + 
    ggtitle(capture[i]) + theme_Publication()
}

umap_plt[[1]] + umap_plt[[2]] + umap_plt[[3]] +
  umap_plt[[4]]
```

## Cell cycle

```{r umap_cc, fig.height=8, fig.width=10}
umap_plt <- list()

for (i in 1:length(sce_list)) {
  current <- sce_list[[i]]
  umap_plt[[i]] <- plotReducedDim(current, dimred="UMAP", colour_by="MKI67") + 
    ggtitle(capture[i]) + theme_Publication()
}

umap_plt[[1]] + umap_plt[[2]] + umap_plt[[3]] +
  umap_plt[[4]]
```

# Visualise UMAP by multiplexing tag calls {.tabset}

## Seurat HTO demux

```{r htodemux, fig.asp=0.8}
sce_tag <- sce_list[c(1:3)]

umap_plt <- list()

for (i in 1:length(sce_tag)) {
  current <- sce_tag[[i]]
  umap_plt[[i]] <- plotReducedDim(current, dimred="UMAP", colour_by="htodemux") + 
  ggtitle(capture[i]) +
  theme_Publication()
}

umap_plt[[1]] + theme(legend.text = element_text(size = 12), plot.title = element_text(size = 14), legend.title=element_blank()) + 
  umap_plt[[2]] + theme(legend.text = element_text(size = 12), plot.title = element_text(size = 14), legend.title=element_blank()) +
  umap_plt[[3]] + theme(legend.text = element_text(size = 12), plot.title = element_text(size = 14), legend.title=element_blank()) +
  plot_layout(ncol=2)
```

## BFF cluster

```{r bff_cluster, fig.height=8, fig.width=10}
umap_plt <- list()

for (i in 1:length(sce_tag)) {
  current <- sce_tag[[i]]
  umap_plt[[i]] <- plotReducedDim(current, dimred="UMAP", colour_by="bff_cluster") + 
  ggtitle(capture[i]) +
  theme_Publication()
}

umap_plt[[1]] + umap_plt[[2]] + umap_plt[[3]] + plot_layout(ncol=2)
```

## GMM demux

```{r gmm_demux, fig.height=8, fig.width=10}
umap_plt <- list()

for (i in 1:length(sce_tag)) {
  current <- sce_tag[[i]]
  umap_plt[[i]] <- plotReducedDim(current, dimred="UMAP", colour_by="gmm_demux") + 
  ggtitle(capture[i]) +
  theme_Publication()
}

umap_plt[[1]] + umap_plt[[2]] + umap_plt[[3]] + plot_layout(ncol=2)
```

## DropletUtils

```{r dropletutils, fig.height=8, fig.width=10}
umap_plt <- list()

for (i in 1:length(sce_tag)) {
  current <- sce_tag[[i]]
  umap_plt[[i]] <- plotReducedDim(current, dimred="UMAP", colour_by="dropletutils") + 
  ggtitle(capture[i]) +
  theme_Publication()
}

umap_plt[[1]] + umap_plt[[2]] + umap_plt[[3]] + plot_layout(ncol=2)
```

# Visualise UMAP by Vireo donor calls


```{r vireo, fig.height=8, fig.width=10}
umap_plt <- list()

for (i in 1:length(sce_tag)) {
  current <- sce_tag[[i]]
  umap_plt[[i]] <- plotReducedDim(current, dimred="UMAP", colour_by="donor_id") + 
  ggtitle(capture[i]) +
  theme_Publication()
}

umap_plt[[1]] + umap_plt[[2]] + umap_plt[[3]] + plot_layout(ncol=2)
```

# Save SCE object

```{r saveSCEs}
saveRDS(sce_list, here::here(
   "/vast/scratch/users/brown.d/S000322/S000322.cluster.rds"
))

saveRDS(sce_list, here::here(
   "/stornext/Projects/score/Analyses/R010_multiplexing/SCEs/PDX/S000322_allcaptures.cluster.rds"
))
```

## Session Info

```{r}
sessionInfo()
```
